Begin implementing support for complex documentation.
Documentation is hard, especially the kind that's self-referential.
Break the documentation expansion into multiple passes: collection,
cross-reference expansion, and simple expansion to begin with.  Define
entities for documentable attributes and methods, and match them with
corresponding documentation.  Begin validating documentation coverage.
Begin extensively documenting the project with its own markup, as a
sanity check for the language and concepts.  Update the command line
rendering tool so it may better convert entire documentation trees in
the future.