Begin inheriting documentation from base classes and roles.
Subclasses often acquire implementation from base classes.  This
change adds the start of documentation inheritance as well.  If a
subclass inherits something but doesn't document or skip it, then
Pod::Plexus will look for the documentation at the point of
implementation.

I also improved the POD coverage checking, and so I had to document
everything before continuing.  The docs are sketchy for now; things
are still changing.