Begin yet another major refactoring.
Pod::Plexus::Document was becoming unwieldy with data and logic for
both the documentation and code sides of a module.  Separate them into
sub-modules.  Make the doc section classes smaller and more granular
so that more common logic is shared in base classes.  This seems to be
working well: It was trivial to add =before, =after and =inherits.