Allow =head1 sections to be =include'd.
There's a problem with this.  Included sections don't look exactly
like ones that were hand-rolled.  They kind of need to for validation
purposes.  Validation might need to be pushed farther down the process
to a point after documentation is rendered but before it's returned.
That's woefully late in the game, though.